Refactor:
- move sora.common to sora.crypto.common
- move enums to sora.crypto.type package
- move all classes related to merkle tree to sora.crypto.merkle package
- move service to sora.crypto.json
- move service.impl to sora.crypto.json
- run optimize import on all files
