+++ src/main/java/jp/co/soramitsu/sora/crypto/json/FlattenException.java
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.sora.json;
+package jp.co.soramitsu.sora.crypto.json;
 
 public class FlattenException extends RuntimeException {
