+++ src/main/java/jp/co/soramitsu/sora/crypto/json/JSONSigner.java
@@ -1,4 +1,4 @@
-package jp.co.soramitsu.sora.crypto.service;
+package jp.co.soramitsu.sora.crypto.json;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.IOException;
