+++ src/main/java/jp/co/soramitsu/sora/crypto/json/NotJsonObjectException.java
@@ -1,11 +1,11 @@
-package jp.co.soramitsu.sora.json;
+package jp.co.soramitsu.sora.crypto.json;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
 public class NotJsonObjectException extends RuntimeException {
 
   public NotJsonObjectException(JsonNode root) {
-    super("Not an object JSON: " + root.toString());
+    super("Not a JSON object :" + root.toString());
   }
 
 }
