+++ src/main/java/jp/co/soramitsu/sora/crypto/type/KeyTypeEnum.java
@@ -1,7 +1,8 @@
-package jp.co.soramitsu.sora.crypto.common;
+package jp.co.soramitsu.sora.crypto.type;
 
 import jp.co.soramitsu.crypto.ed25519.EdDSAKey;
 import jp.co.soramitsu.crypto.ed25519.EdDSASecurityProvider;
+import jp.co.soramitsu.sora.crypto.common.Consts;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
