+++ src/test/groovy/jp/co/soramitsu/sora/crypto/service/impl/JSONVerifierImplTest.groovy
@@ -3,16 +3,16 @@ package jp.co.soramitsu.sora.crypto.service.impl
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.node.ObjectNode
 import jp.co.soramitsu.sora.common.MockSignature
-import jp.co.soramitsu.sora.crypto.common.SignatureTypeEnum
+import jp.co.soramitsu.sora.crypto.json.JSONCanonizer
+import jp.co.soramitsu.sora.crypto.json.JSONVerifier
+import jp.co.soramitsu.sora.crypto.json.JSONVerifierImpl
 import jp.co.soramitsu.sora.crypto.proof.Options
 import jp.co.soramitsu.sora.crypto.proof.Proof
-import jp.co.soramitsu.sora.crypto.service.JSONCanonizer
-import jp.co.soramitsu.sora.crypto.service.JSONVerifier
+import jp.co.soramitsu.sora.crypto.type.SignatureTypeEnum
 import spock.lang.Specification
 
 import java.security.PublicKey
 import java.security.Signature
-
 
 class JSONVerifierImplTest extends Specification {
