+++ src/test/groovy/jp/co/soramitsu/sora/json/SaltifierTest.groovy
@@ -2,6 +2,7 @@ package jp.co.soramitsu.sora.json
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import jp.co.soramitsu.sora.common.SaltGenerator
+import jp.co.soramitsu.sora.crypto.json.Saltifier
 import spock.lang.Specification
 
 class SaltifierTest extends Specification {
